requirements.txt
pypdf=2.0.0